Revive swapping keys j<->gj and k<->gk
With the default key bindings, we need to type gj/gk many times to move
around logical lines, and we need to type j/k once to just skip a very
long line.  On the contrary, with the swapped bindings, we can just
press and hold j/k to scroll wrapped lines, and we only need to type
gj/gk once to skip such a line.