Uninstall vim-asciidoc
The dependency 'VimRegStyle' defines mappings prefixed with <C-b> in
insert mode which introduce wait after pressing <C-b> in insert mode.
Because of the wait, it is very difficult to see where the cursor is
when we move the cursor in insert mode by <C-b>.
It seems that VimRegStyle doesn't provide an option to stop it from
defining those mappings, so I decided to uninstall it.