Fix an issue that yy doesn't yank immediately
This was caused by the fact that `map` define a key mapping also in the
operator mode.  When we press `y` once, the operator mode begins.
Therefore, the next input of `y` will be ambiguous because there is a
mapping `ys` in the operator mode, and vim will waits until `s` is
pressed or it reaches a timeout.