feat: Add colorscheme 'folke/tokyonight.nvim', also abandon lazy loading non-default colorschemes
The tokyonight can accept `opts`.  It is difficult to support both `opts` and
`config` because we cannot rely on lazy.nvim's default implementation of
`config` function that runs `require(MAIN).setup(opts)` appropriately.
To utilize the default implementation, we cannot define our own `config` and
will not be able to call `colorscheme` command.